chore(shared): move functions/ to functions/legacy
Current functions/ folder contains logic to unify syncs and on-events.
We are going to introduce data access layers and services to manage
native functions backed by functions_configs, functions_instances, ... tables.
This commit is first moving the legacy unification into a `legacy/`
subfolder so native functions logic can live at the root.
