+++ packages/shared/lib/services/functions/legacy/index.ts
@@ -1,0 +1,3 @@
+// Legacy unification over `_nango_sync_configs` and `on_event_scripts`.
+export { getFunction, listFunctions } from './service.js';
+export { findActiveDeployedMeta } from './models/functions.js';
