+++ packages/shared/lib/services/functions/legacy/mappers.ts
@@ -1,36 +1,16 @@
 import { Err, Ok } from '@nangohq/utils';
 
+import { eventTypeMapper } from '../../on-event-scripts.service.js';
+
 import type { FunctionRow } from './models/functions.js';
 import type {
     DBOnEventScript,
     DeployedNangoActionFunction,
     DeployedNangoFunction,
     DeployedNangoOnEventFunction,
-    DeployedNangoSyncFunction,
-    OnEventType
+    DeployedNangoSyncFunction
 } from '@nangohq/types';
 import type { Result } from '@nangohq/utils';
-
-const DB_TO_API_EVENT_TYPE: Record<DBOnEventScript['event'], OnEventType> = {
-    POST_CONNECTION_CREATION: 'post-connection-creation',
-    PRE_CONNECTION_DELETION: 'pre-connection-deletion',
-    VALIDATE_CONNECTION: 'validate-connection'
-} as const;
-
-const API_TO_DB_EVENT_TYPE: Record<OnEventType, DBOnEventScript['event']> = {
-    'post-connection-creation': 'POST_CONNECTION_CREATION',
-    'pre-connection-deletion': 'PRE_CONNECTION_DELETION',
-    'validate-connection': 'VALIDATE_CONNECTION'
-} as const;
-
-export const eventTypeMapper = {
-    fromDb: (event: DBOnEventScript['event']): OnEventType => {
-        return DB_TO_API_EVENT_TYPE[event];
-    },
-    toDb: (eventType: OnEventType): DBOnEventScript['event'] => {
-        return API_TO_DB_EVENT_TYPE[eventType];
-    }
-};
 
 export function toDeployedNangoFunction(row: FunctionRow): Result<DeployedNangoFunction> {
     const description = row.metadata?.description;
@@ -67,7 +47,7 @@ export function toDeployedNangoFunction(row: FunctionRow): Result<DeployedNangoF
                 return Err(new Error('Unknown on-event type: null'));
             }
 
-            const apiEvent = DB_TO_API_EVENT_TYPE[row.event as DBOnEventScript['event']];
+            const apiEvent = eventTypeMapper.fromDb(row.event as DBOnEventScript['event']);
             if (!apiEvent) {
                 return Err(new Error(`Unknown on-event type: ${row.event}`));
             }
